requirements.txt
# Telegram Bot
python-telegram-bot[all]==21.0

# Yandex.Music
yandex-music==2.2.0

# HTTP Client
httpx==0.27.0

# Data Validation
pydantic==2.6.0
pydantic-settings==2.1.0

# Configuration
python-dotenv==1.0.1

# Caching
redis==5.0.1

# Development
pytest==7.4.4
pytest-asyncio==0.23.4
pytest-cov==4.1.0
black==24.1.1
ruff==0.2.1
mypy==1.8.0

# Logging
loguru==0.7.2
